defconfigs: rename, disable keyboard dock
This driver always reports itself as present even when the keyboard is
not.  This causes the software keyboard to be disabled.  Almost no one
has this accessory anyway - so disable it until someone who actually has
a keyboard to test with can fix this broken driver. "#
CONFIG_KEYBOARD_SEC_DOCK is not set"
---
defconfig renamed to custom